Installing and configuring capybara tests

+++ db/migrate/20190702152855_create_champion.rb
@@ -17,7 +17,7 @@ class CreateChampion < ActiveRecord::Migration[5.2]
       t.integer :starting_mana, null: false
       t.integer :mana_cost, null: false
       t.integer :attack_damage, null: false
-      t.integer :attack_speed, null: false
+      t.float :attack_speed, null: false
       t.integer :attack_range, null: false
       t.integer :armor, null: false
       t.integer :magic_resist, null: false
